fixed routes for login (put them all in main). Can now log in and log out.
